Seperate detectors to different files for android
