+++ packages/firebase_ml_vision/android/src/main/java/io/flutter/plugins/firebasemlvision/FaceDetector.java
@@ -1,0 +1,4 @@
+package io.flutter.plugins.firebasemlvision;
+
+public class FaceDetector {
+}
