+++ packages/firebase_ml_vision/android/src/main/java/io/flutter/plugins/firebasemlvision/FirebaseMlVisionPlugin.java
@@ -1,31 +1,21 @@
 package io.flutter.plugins.firebasemlvision;
 
-import android.graphics.Point;
-import android.graphics.Rect;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.ml.vision.FirebaseVision;
+
 import com.google.firebase.ml.vision.common.FirebaseVisionImage;
-import com.google.firebase.ml.vision.text.FirebaseVisionText;
-import com.google.firebase.ml.vision.text.FirebaseVisionTextDetector;
+
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
+
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 /** FirebaseMlVisionPlugin */
 public class FirebaseMlVisionPlugin implements MethodCallHandler {
   private Registrar registrar;
-  private FirebaseVisionTextDetector textDetector;
 
   private FirebaseMlVisionPlugin(Registrar registrar) {
     this.registrar = registrar;
@@ -54,18 +44,19 @@ public class FirebaseMlVisionPlugin implements MethodCallHandler {
       case "LabelDetector#close":
         break;
       case "TextDetector#detectInImage":
-        handleTextDetectionResult(call, result);
+        FirebaseVisionImage image = filePathToVisionImage((String) call.arguments, result);
+        if (image != null) TextDetector.handleTextDetectionResult(image, result);
         break;
       case "TextDetector#close":
-        if (textDetector != null) {
+        if (TextDetector.textDetector != null) {
           try {
-            textDetector.close();
+            TextDetector.textDetector.close();
             result.success(null);
           } catch (IOException exception) {
             result.error("textDetectorError", exception.getLocalizedMessage(), null);
           }
 
-          textDetector = null;
+          TextDetector.textDetector = null;
         }
         break;
       default:
@@ -73,88 +64,14 @@ public class FirebaseMlVisionPlugin implements MethodCallHandler {
     }
   }
 
-  /////////////// Barcode Detector ///////////////
+  private FirebaseVisionImage filePathToVisionImage(String path, Result result) {
+    File file = new File(path);
 
-  /////////////// Face Detector ///////////////
-
-  /////////////// Label Detector ///////////////
-
-  /////////////// Text Detector ///////////////
-  private void handleTextDetectionResult(MethodCall call, final Result result) {
-    File file = new File((String) call.arguments);
-
-    FirebaseVisionImage image;
     try {
-      image = FirebaseVisionImage.fromFilePath(registrar.context(), Uri.fromFile(file));
+      return FirebaseVisionImage.fromFilePath(registrar.context(), Uri.fromFile(file));
     } catch (IOException exception) {
-      result.error("textDetectorError", exception.getLocalizedMessage(), null);
-      return;
+      result.error("textDetectorIOError", exception.getLocalizedMessage(), null);
+      return null;
     }
-
-    if (textDetector == null) textDetector = FirebaseVision.getInstance().getVisionTextDetector();
-    textDetector
-        .detectInImage(image)
-        .addOnSuccessListener(
-            new OnSuccessListener<FirebaseVisionText>() {
-              @Override
-              public void onSuccess(FirebaseVisionText firebaseVisionText) {
-                List<Map<String, Object>> blocks = new ArrayList<>();
-                for (FirebaseVisionText.Block block : firebaseVisionText.getBlocks()) {
-                  Map<String, Object> blockData = new HashMap<>();
-                  addTextData(
-                      blockData, block.getBoundingBox(), block.getCornerPoints(), block.getText());
-
-                  List<Map<String, Object>> lines = new ArrayList<>();
-                  for (FirebaseVisionText.Line line : block.getLines()) {
-                    Map<String, Object> lineData = new HashMap<>();
-                    addTextData(
-                        lineData, line.getBoundingBox(), line.getCornerPoints(), line.getText());
-
-                    List<Map<String, Object>> elements = new ArrayList<>();
-                    for (FirebaseVisionText.Element element : line.getElements()) {
-                      Map<String, Object> elementData = new HashMap<>();
-                      addTextData(
-                          elementData,
-                          element.getBoundingBox(),
-                          element.getCornerPoints(),
-                          element.getText());
-                      elements.add(elementData);
-                    }
-                    lineData.put("elements", elements);
-                    lines.add(lineData);
-                  }
-                  blockData.put("lines", lines);
-                  blocks.add(blockData);
-                }
-                result.success(blocks);
-              }
-            })
-        .addOnFailureListener(
-            new OnFailureListener() {
-              @Override
-              public void onFailure(@NonNull Exception exception) {
-                result.error("textDetectorError", exception.getLocalizedMessage(), null);
-              }
-            });
-  }
-
-  private void addTextData(
-      Map<String, Object> addTo, Rect boundingBox, Point[] cornerPoints, String text) {
-    addTo.put("text", text);
-
-    if (boundingBox != null) {
-      addTo.put("left", boundingBox.left);
-      addTo.put("top", boundingBox.top);
-      addTo.put("width", boundingBox.width());
-      addTo.put("height", boundingBox.height());
-    }
-
-    List<int[]> points = new ArrayList<>();
-    if (cornerPoints != null) {
-      for (Point point : cornerPoints) {
-        points.add(new int[] {point.x, point.y});
-      }
-    }
-    addTo.put("points", points);
   }
 }
