+++ packages/firebase_ml_vision/android/src/main/java/io/flutter/plugins/firebasemlvision/LabelDetector.java
@@ -1,0 +1,4 @@
+package io.flutter.plugins.firebasemlvision;
+
+public class LabelDetector {
+}
